+++ cpu/efm32/periph/timer.c
@@ -73,7 +73,7 @@ static inline bool _is_letimer(timer_t dev)
 #endif
 }
 
-static void _letimer_init(tim_t dev, unsigned long freq)
+static void _letimer_init(tim_t dev, uint32_t freq)
 {
     (void) freq;
 #if LETIMER_COUNT
@@ -98,7 +98,7 @@ static void _letimer_init(tim_t dev, unsigned long freq)
 #endif
 }
 
-static void _timer_init(tim_t dev, unsigned long freq)
+static void _timer_init(tim_t dev, uint32_t freq)
 {
     TIMER_TypeDef *pre, *tim;
 
@@ -143,7 +143,7 @@ static void _timer_init(tim_t dev, unsigned long freq)
     TIMER_IntEnable(tim, TIMER_IEN_CC0 | TIMER_IEN_CC1 | TIMER_IEN_CC2);
 }
 
-int timer_init(tim_t dev, unsigned long freq, timer_cb_t callback, void *arg)
+int timer_init(tim_t dev, uint32_t freq, timer_cb_t callback, void *arg)
 {
     /* test if given timer device is valid */
     if (dev >= TIMER_NUMOF) {
