+++ cpu/msp430fxyz/periph/timer.c
@@ -41,7 +41,7 @@ static timer_cb_t isr_cb;
 static void *isr_arg;
 
 
-int timer_init(tim_t dev, unsigned long freq, timer_cb_t cb, void *arg)
+int timer_init(tim_t dev, uint32_t freq, timer_cb_t cb, void *arg)
 {
     /* using fixed TIMER_BASE for now */
     if (dev != 0) {
